+++ CHANGELOG.md
@@ -248,7 +248,7 @@
 
 - Shopify now requires Craft CMS 5.0.0-beta.10 or later.
 
-## 4.1.2 - 2024-04-15 
+## 4.1.2 - 2024-04-15
 
 - Fixed a PHP error that could occur when syncing products with emojis. ([#107](https://github.com/craftcms/shopify/issues/107))
 - Fixed a PHP error that could occur when syncing products. ([#105](https://github.com/craftcms/shopify/issues/105))